requirements.txt
Django==1.5.4
MySQL-python==1.2.4
PyYAML==3.10
Werkzeug==0.9.4
amqp==1.0.13
anyjson==0.3.3
argh==0.23.3
billiard==2.7.3.34
celery==3.0.24
django-celery==3.0.23
django-kombu==0.9.4
django-supervisor==0.3.1
httplib2==0.8
kombu==2.5.15
meld3==0.6.10
pathtools==0.1.2
python-dateutil==2.1
pytz==2013.7
six==1.4.1
supervisor==3.0
twilio==3.6.2
unittest2==0.5.1
watchdog==0.6.0
wsgiref==0.1.2
django-tastypie==0.11.0
